+++ apps/client/src/environments/patch-notes.ts
@@ -1,27 +1,6 @@
 export const patchNotes = `### Bug Fixes
 
-* **alarms:** fixed duplicate map names in custom alarm popup.
-* **db:** fixed "Can contain items" container overflowing.
-* **db:** fixed lvl 10 la noscea node having wrong Y coord.
-* **db:** fixed missing db links on some shops.
-* **db:** fixed wrong collectable flags on some items.
-* **desktop:** fix for slow autofill with non-gathering/crafting obtention.
-* **desktop:** fixed some pages not loading properly when opening the app.
-* **list:** no longer show skip tag if skippable amount is <= 0.
-* **user:** fixed free company ID not applied properly to user profile.
-
-
-### Features
-
-* **allagan-reports:** added id in [] to avoid confusion, more details on drops as well.
-* **currency-spending:** added Voidrakes and Althyk Lavender.
-* **db:** added monster location to the quick search when known.
-* **inventory-optimizer:** new optimizer for items that are now functioning as NQ.
-* **search:** added gathering node to quick search.
-* **simulator:** new error message for when level or stats are too low for the recipe.
-
-
-### Performance Improvements
-
-* **lists:** improved list details loading performances.
-* **lists:** migrated the history to its own subcollection.`;
+* **gearsets:** fixed etro gearset import with identical rings.
+* **inventory-optimizer:** collectables won't be listed as deprecated HQ anymore.
+* **lists:** fixed lists page sometimes not loading properly.
+* **pricing:** prices are now properly populated from vendor prices.`;